requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-ace==1.0.2
django-admin-bootstrapped==2.5.6
django-admin-tools==0.6.0
django-ckeditor==5.0.2
django-easy-pjax==1.2.0
django-filer==0.9.12
django-grappelli==2.6.3
django-material==0.4.1
django-mptt==0.7.4
django-polymorphic==0.7.2
django-suit==0.2.15
django-toolbelt==0.0.1
django-widget-tweaks==1.3
easy-thumbnails==2.2
gunicorn==19.3.0
MySQL-python==1.2.5
oauthlib==1.0.3
Pillow==2.7.0
python-dateutil==2.4.2
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
Unidecode==0.4.18
wheel==0.24.0
django-imagekit==3.3
